requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
requests==2.32.3
Jinja2==3.1.4
aiofiles==24.1.0
python-multipartpython-dotenv
